Take logger out in utils, as used in multiple places
